# api/openapi.yaml
openapi: 3.0.3
info:
  title: TradingCopilot API
  version: 0.1.0
  description: JSON:API contract target for the Go modular monolith migration.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /auth/bootstrap:
    post:
      operationId: postAuthBootstrap
      security: []
      requestBody: { $ref: "#/components/requestBodies/AuthCredentialsRequest" }
      responses:
        "200": { $ref: "#/components/responses/AuthTokenDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
        "409": { $ref: "#/components/responses/JsonApiErrorDocument" }
  /auth/login:
    post:
      operationId: postAuthLogin
      security: []
      requestBody: { $ref: "#/components/requestBodies/AuthCredentialsRequest" }
      responses:
        "200": { $ref: "#/components/responses/AuthTokenDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
        "401": { $ref: "#/components/responses/JsonApiErrorDocument" }
  /auth/logout:
    post:
      operationId: postAuthLogout
      responses:
        "200": { $ref: "#/components/responses/AuthSessionDocument" }
        "401": { $ref: "#/components/responses/JsonApiErrorDocument" }
...
  /audit-events:
    get:
      operationId: getAuditEvents
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/AuditEventCollectionDocument" }
...
  /ops/jobs:
    get:
      operationId: getOpsJobs
      parameters:
        - $ref: "#/components/parameters/OpsQueueName"
        - $ref: "#/components/parameters/OpsTaskType"
        - $ref: "#/components/parameters/OpsTaskState"
        - $ref: "#/components/parameters/OpsFailedLimit"
      responses:
        "200": { $ref: "#/components/responses/OpsJobsDocument" }
  /ops/jobs/retry-failed:
    post:
      operationId: postOpsJobsRetryFailed
      responses:
        "200": { $ref: "#/components/responses/OpsJobActionDocument" }
        "409": { $ref: "#/components/responses/JsonApiErrorDocument" }
...
  /message-subscriptions:
    get:
      operationId: getMessageSubscriptions
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionCollectionDocument" }
    post:
      operationId: postMessageSubscriptions
      requestBody: { $ref: "#/components/requestBodies/MessageSubscriptionUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionDocument" }
...
  /message-subscription-filters:
    get:
      operationId: getMessageSubscriptionFilters
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionFilterCollectionDocument" }
    post:
      operationId: postMessageSubscriptionFilters
      requestBody: { $ref: "#/components/requestBodies/MessageSubscriptionFilterUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/MessageSubscriptionFilterDocument" }
...
  /ingested-messages:
    get:
      operationId: getIngestedMessages
      parameters:
        - $ref: "#/components/parameters/SubscriptionIdQuery"
        - $ref: "#/components/parameters/ResearchTeamId"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/OnlyUnfiltered"
        - $ref: "#/components/parameters/FilterDecision"
        - $ref: "#/components/parameters/FilterStatus"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/IngestedMessageCollectionDocument" }
    post:
      operationId: postIngestedMessages
      requestBody: { $ref: "#/components/requestBodies/IngestedMessageUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/IngestedMessageDocument" }
...
  /meetings:
    get:
      operationId: getMeetings
      parameters:
        - $ref: "#/components/parameters/ResearchTeamId"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/TriggerSource"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/MeetingCollectionDocument" }
    post:
      operationId: postMeetings
      requestBody: { $ref: "#/components/requestBodies/MeetingStartRequest" }
      responses:
        "200": { $ref: "#/components/responses/MeetingDocument" }
...
  /wake-plans:
    get:
      operationId: getWakePlans
      parameters:
        - $ref: "#/components/parameters/ResearchTeamId"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/MeetingIdQuery"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/WakePlanCollectionDocument" }
    post:
      operationId: postWakePlans
      requestBody: { $ref: "#/components/requestBodies/WakePlanUpsertRequest" }
      responses:
        "200": { $ref: "#/components/responses/WakePlanDocument" }
        "400": { $ref: "#/components/responses/JsonApiErrorDocument" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    AuthCredentialsRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/AuthCredentialsDocument"
...
    WakePlanUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanUpsertDocument"
...
    MessageSubscriptionUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionUpsertDocument"
    MessageSubscriptionFilterUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionFilterUpsertDocument"
...
    IngestedMessageUpsertRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/IngestedMessageUpsertDocument"
...
    MeetingStartRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingStartDocument"
...
  responses:
...
    AuthTokenDocument:
      description: Auth token JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/AuthTokenDocument"
...
    AuthSessionDocument:
      description: Auth session JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/AuthSessionDocument"
    AuditEventCollectionDocument:
      description: Audit event JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/AuditEventCollectionDocument"
...
    OpsJobsDocument:
      description: Ops jobs JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/OpsJobsDocument"
    OpsJobActionDocument:
      description: Ops job action JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/OpsJobActionDocument"
...
    JsonApiErrorDocument:
      description: JSON:API error document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/JsonApiErrorDocument"
...
    WakePlanCollectionDocument:
      description: Wake plan JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanCollectionDocument"
    WakePlanDocument:
      description: Wake plan JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/WakePlanDocument"
    MeetingCollectionDocument:
      description: Meeting JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingCollectionDocument"
    MeetingDocument:
      description: Meeting JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingDocument"
...
    MessageSubscriptionCollectionDocument:
      description: Message subscription JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionCollectionDocument"
...
    MessageSubscriptionDocument:
      description: Message subscription JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionDocument"
    MessageSubscriptionFilterCollectionDocument:
      description: Message subscription filter JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionFilterCollectionDocument"
    MessageSubscriptionFilterDocument:
      description: Message subscription filter JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MessageSubscriptionFilterDocument"
...
    IngestedMessageCollectionDocument:
      description: Ingested message JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/IngestedMessageCollectionDocument"
    IngestedMessageDocument:
      description: Ingested message JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/IngestedMessageDocument"
...
  parameters:
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
...
    Status:
      name: status
      in: query
      schema: { type: string }
    TriggerSource:
      name: triggerSource
      in: query
      schema: { type: string }
    Tag:
      name: tag
      in: query
      schema: { type: string }
    MeetingIdQuery:
      name: meetingId
      in: query
      schema: { type: string }
    ResearchTeamId:
      name: researchTeamId
      in: query
      schema: { type: string }
...
    SubscriptionIdQuery:
      name: subscriptionId
      in: query
      schema: { type: string }
...
    OnlyUnfiltered:
      name: onlyUnfiltered
      in: query
      schema: { type: boolean }
    FilterDecision:
      name: filterDecision
      in: query
      schema: { type: string }
    FilterStatus:
      name: filterStatus
      in: query
      schema: { type: string }
...
    OpsQueueName:
      name: queue
      in: query
      description: Optional Redis/asynq queue name used to filter failed task samples.
      schema: { type: string }
    OpsTaskType:
      name: type
      in: query
      description: Optional task type used to filter failed task samples.
      schema: { type: string }
    OpsTaskState:
      name: state
      in: query
      description: Optional failed task state used to filter samples.
      schema:
        type: string
        enum: [retry, archived]
    OpsFailedLimit:
      name: failedLimit
      in: query
      description: Maximum number of failed task samples to return. Values above 50 are capped.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 8
    PageLimit:
      name: page[limit]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    PageCursor:
      name: page[cursor]
      in: query
      schema:
        type: string
...
  schemas:
...
    JsonApiErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ErrorObject"
    AuthCredentialsDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthCredentialsResource"
    AuthCredentialsResource:
      type: object
      required: [type, attributes]
      properties:
        type:
          type: string
          enum: [auth-login-requests, auth-bootstrap-requests]
        attributes:
          $ref: "#/components/schemas/AuthCredentialsAttributes"
...
    AuthTokenDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthTokenResource"
    AuthTokenResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [auth-tokens]
        id:
          type: string
          enum: [current]
        attributes:
          $ref: "#/components/schemas/AuthTokenAttributes"
...
    AuthSessionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthSessionResource"
    AuthSessionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [auth-sessions]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/AuthSessionAttributes"
...
    AuditEventCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditEventResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    AuditEventResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [audit-events]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/AuditEventAttributes"
...
    OpsJobsDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/OpsJobsResource"
    OpsJobsResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [ops-job-reports]
        id:
          type: string
          enum: [current]
        attributes:
          type: object
          additionalProperties: true
          properties:
            generatedAt: { type: string, format: date-time }
            meetingDispatchMode: { type: string }
            paperExecutionMode: { type: string }
            queue:
              $ref: "#/components/schemas/OpsQueueDiagnostics"
            recentErrors:
              $ref: "#/components/schemas/OpsRecentErrors"
            redis:
              type: object
              additionalProperties: true
            worker:
              type: object
              additionalProperties: true
            scheduler:
              type: object
              additionalProperties: true
            messageListener:
              type: object
              additionalProperties: true
            paperEngine:
              type: object
              additionalProperties: true
    OpsJobActionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/OpsJobActionResource"
    OpsJobActionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [ops-job-actions]
        id:
          type: string
          enum: [retry-failed]
        attributes:
          type: object
          additionalProperties: true
          properties:
            status: { type: string }
            action: { type: string }
            retriedRetryCount: { type: integer }
            retriedArchiveCount: { type: integer }
            totalSubmitted: { type: integer }
            queues:
              type: array
              items:
                type: object
                additionalProperties: true
            ranAt: { type: string, format: date-time }
            destructive: { type: boolean }
            detail: { type: string }
...
    WakePlanCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WakePlanResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    WakePlanDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/WakePlanResource"
    WakePlanResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [wake-plans]
        id:
          type: string
        attributes:
          $ref: "#/components/schemas/WakePlanAttributes"
        relationships:
          type: object
          properties:
            meeting:
              $ref: "#/components/schemas/Relationship"
            sourceMeetingEvent:
              $ref: "#/components/schemas/Relationship"
...
    WakePlanUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type:
              type: string
              enum: [wake-plans]
            id:
              type: string
            attributes:
              $ref: "#/components/schemas/WakePlanInputAttributes"
...
    MeetingCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MeetingResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    MeetingDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MeetingResource"
    MeetingResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [meetings] }
        id: { type: string }
        attributes:
          $ref: "#/components/schemas/MeetingAttributes"
...
    MeetingStartDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [meetings] }
            attributes:
              type: object
              required: [researchTeamId, topic]
              properties:
                researchTeamId: { type: integer }
                topic: { type: string }
                triggerSource: { type: string }
                context:
                  type: object
                  additionalProperties: true
...
    MessageSubscriptionCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageSubscriptionResource"
    MessageSubscriptionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MessageSubscriptionResource"
...
    MessageSubscriptionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [message-subscriptions] }
        id: { type: string }
        attributes:
          type: object
          properties:
            provider:
              type: string
              enum: [telegram_channel, rss_feed]
            title: { type: string }
            sourceRef:
              type: string
              description: RSS/Atom source refs must be http or https feed URLs without URL-embedded credentials; use rssAuthType, rssUsername, and rssPassword for private feeds.
            enabled: { type: boolean }
            filterId: { type: integer }
            filterName: { type: string }
            teamIds:
              type: array
              items: { type: integer }
            backfillLimit: { type: integer }
            pollIntervalSeconds: { type: integer }
            collectFrom: { type: string, format: date-time }
            lastCollectedAt:
              type: string
              format: date-time
              nullable: true
            nextCollectAt:
              type: string
              format: date-time
              nullable: true
            lastCollectError:
              type: string
              nullable: true
            config:
              type: object
              additionalProperties: true
            rssAuthType:
              type: string
              enum: [none, basic, bearer]
            rssUsername: { type: string }
            hasRssPassword: { type: boolean }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    MessageSubscriptionUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [message-subscriptions] }
            id: { type: string }
            attributes:
              type: object
              required: [title, sourceRef]
              properties:
                provider:
                  type: string
                  enum: [telegram_channel, rss_feed]
                title: { type: string }
                sourceRef:
                  type: string
                  description: RSS/Atom source refs must be http or https feed URLs without URL-embedded credentials; use rssAuthType, rssUsername, and rssPassword for private feeds.
                enabled: { type: boolean }
                filterId: { type: integer }
                teamIds:
                  type: array
                  items: { type: integer }
                backfillLimit: { type: integer }
                pollIntervalSeconds: { type: integer }
                config:
                  type: object
                  additionalProperties: true
                rssAuthType:
                  type: string
                  enum: [none, basic, bearer]
                rssUsername:
                  type: string
                  description: Username for RSS Basic auth. Ignored for Bearer unless a provider needs a label.
                rssPassword:
                  type: string
                  description: RSS Basic password or Bearer token. Stored as an encrypted secret and never returned.
    MessageSubscriptionFilterCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MessageSubscriptionFilterResource"
    MessageSubscriptionFilterDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MessageSubscriptionFilterResource"
    MessageSubscriptionFilterResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [message-subscription-filters] }
        id: { type: string }
        attributes:
          $ref: "#/components/schemas/MessageSubscriptionFilterAttributes"
        relationships:
          type: object
          properties:
            provider:
              $ref: "#/components/schemas/Relationship"
...
    MessageSubscriptionFilterUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [message-subscription-filters] }
            id: { type: string }
            attributes:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                promptTemplate: { type: string }
                providerId:
                  type: integer
                  nullable: true
                model:
                  type: string
                  nullable: true
                enabled: { type: boolean }
                isDefault: { type: boolean }
...
    IngestedMessageCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/IngestedMessageResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    IngestedMessageDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/IngestedMessageResource"
    IngestedMessageResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [ingested-messages] }
        id: { type: string }
        attributes:
          $ref: "#/components/schemas/IngestedMessageAttributes"
        relationships:
          type: object
          properties:
            subscription:
              $ref: "#/components/schemas/Relationship"
...
    IngestedMessageUpsertDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [ingested-messages] }
            id: { type: string }
            attributes:
              type: object
              required: [subscriptionId, text]
              properties:
                subscriptionId: { type: integer }
                sourceMessageId: { type: string }
                messageTime: { type: string, format: date-time }
                text: { type: string }
                filterDecision:
                  type: string
                  nullable: true
                filterReason:
                  type: string
                  nullable: true
                relatedSymbols:
                  type: array
                  items: { type: string }
...
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
        field:
          type: string
    Links:
      type: object
      properties:
        self:
          type: string
        next:
          type: string
        prev:
          type: string